+++ docker/bootstrap/bootstrap.sh
@@ -1,0 +1,28 @@
+#!/usr/bin/env bash
+set -e
+
+# Copies any default config files from CassKop bootstrapper image to the /etc/cassandra volume which will replace the one used in cassandra image
+cp -rLv /${BOOTSTRAP_CONF}/* /etc/cassandra/
+
+# If User has submited a configMap, we uses them to deplace default ones
+# (overwriting the above)
+if [[ -d /${CONFIGMAP} && ! -z `ls -A /${CONFIGMAP}` ]] ; then
+    echo "We have a ConfigMap, we surcharge default configuration files"
+    cp -rLv /${CONFIGMAP}/* /etc/cassandra/
+fi
+
+# Copies any extra libraries from this bootstrapper image to the extra-lib empty-dir
+if [[ -d /${BOOTSTRAP_LIBS} && ! -z `ls -A /${BOOTSTRAP_LIBS}` ]] ; then
+    echo "We have a Additional libs, we surcharge default configuration files"
+    cp -v ${BOOTSTRAP_LIBS}/* $CASSANDRA_LIBS/
+   fi
+
+if [ -f ${CONFIGMAP}/pre_run.sh ]; then
+    echo "We found pre_run.sh script, we execute it"
+    ${CONFIGMAP}/pre_run.sh
+fi
+
+# Bootstrap Cassandra configuration
+/${BOOTSTRAP_CONF}/run.sh
+
+echo 'bootstrap ended :-)'